fix: address claude review
