+++ app/src/main/java/to/bitkit/ui/components/CenteredProfileHeader.kt
@@ -35,7 +35,7 @@ fun CenteredProfileHeader(
 ) {
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = modifier,
+        modifier = modifier
     ) {
         Text13Up(
             text = publicKey.ellipsisMiddle(TRUNCATED_PK_LENGTH),
@@ -53,13 +53,13 @@ fun CenteredProfileHeader(
                 modifier = Modifier
                     .size(100.dp)
                     .clip(CircleShape)
-                    .background(Colors.Gray5),
+                    .background(Colors.Gray5)
             ) {
                 Icon(
                     painter = painterResource(R.drawable.ic_user_square),
                     contentDescription = null,
                     tint = Colors.White32,
-                    modifier = Modifier.size(50.dp),
+                    modifier = Modifier.size(50.dp)
                 )
             }
         }
@@ -71,7 +71,7 @@ fun CenteredProfileHeader(
             maxLines = 2,
             overflow = TextOverflow.Ellipsis,
             textAlign = TextAlign.Center,
-            modifier = if (nameTestTag != null) Modifier.testTag(nameTestTag) else Modifier,
+            modifier = if (nameTestTag != null) Modifier.testTag(nameTestTag) else Modifier
         )
 
         if (bio.isNotEmpty()) {
@@ -80,7 +80,7 @@ fun CenteredProfileHeader(
                 text = bio,
                 color = Colors.White64,
                 textAlign = TextAlign.Center,
-                modifier = if (notesTestTag != null) Modifier.testTag(notesTestTag) else Modifier,
+                modifier = if (notesTestTag != null) Modifier.testTag(notesTestTag) else Modifier
             )
         }
     }
