+++ app/src/main/java/to/bitkit/ui/components/Text.kt
@@ -38,7 +38,7 @@ fun Display(
         maxLines = maxLines,
         overflow = overflow,
         textAlign = textAlign,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -53,7 +53,7 @@ fun Display(
         style = AppTextStyles.Display.merge(
             color = color,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -68,7 +68,7 @@ fun Headline(
         style = AppTextStyles.Headline.merge(
             color = color,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -86,7 +86,7 @@ fun Headline20(
             letterSpacing = (-.5).sp,
             color = color,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -103,7 +103,7 @@ fun Headline24(
             lineHeight = 24.sp,
             color = color,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -124,7 +124,7 @@ fun Title(
             color = color,
             textAlign = textAlign,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -145,7 +145,7 @@ fun Subtitle(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -189,7 +189,7 @@ fun BodyM(
         maxLines = maxLines,
         minLines = minLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -229,7 +229,7 @@ fun BodyMSB(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -250,7 +250,7 @@ fun BodyMB(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -343,7 +343,7 @@ fun BodySB(
             color = color,
             textAlign = textAlign,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -360,7 +360,7 @@ fun Text13Up(
             color = color,
             textAlign = textAlign,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -381,7 +381,7 @@ fun Caption(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -438,7 +438,7 @@ fun Caption13Up(
             color = color,
             textAlign = textAlign,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -459,6 +459,6 @@ fun Footnote(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
